backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4
django-cors-headers>=4.3,<5
django-oauth-toolkit>=3.0,<4
social-auth-app-django>=5.4,<6
drf-social-oauth2>=2.1,<3
djangorestframework-simplejwt>=5.3,<6
Pillow>=10.0,<11
django-imagekit>=5.0,<6
django-filter>=24.0,<25
mercadopago>=2.3,<3
